+++ apps/daimo-bot/src/botResponses.ts
