+++ apps/daimo-bot/src/scripts/registerNeynarSigner.ts
