+++ apps/daimo-bot/src/test/index.test.ts
