+++ apps/daimo-bot/src/test/jest.setup.js
