+++ apps/daimo-bot/src/trpcClient.ts
