+++ apps/daimo-bot/src/types.ts
